# openapi.yaml
openapi: 3.1.0
info:
  title: Powerlifting App API
  version: 1.0.0
  description: |
    REST API for the Powerlifting App — a multi-user workout tracking platform
    for athletes and coaches.

    ## Auth
    All endpoints except `/auth/register`, `/auth/login`, and `/auth/token`
    require `Authorization: Bearer <access_token>`.

    ## Errors
    All error responses use **RFC 7807 Problem Details** with
    `Content-Type: application/problem+json`.

    ## REQ mapping
    Each operation carries an `x-req` extension listing the requirement IDs it
    fulfils (e.g. `x-req: [REQ-001]`).
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# =============================================================================
# COMPONENTS
# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    ISODateTime:
      type: string
      format: date-time
      example: "2026-03-11T14:00:00Z"
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: "https://api.example.com/problems/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "email must be a valid email address"
        instance:
          type: string
          format: uri
          example: "/api/v1/auth/register"

    # --- Auth (REQ-001, REQ-002) ----------------------------------------------
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          example: "lifter@example.com"
        password:
          type: string
          minLength: 8
          example: "Str0ngP@ss"
        role:
          type: string
          enum: [athlete, coach]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (15 min TTL). Refresh token returned as httpOnly cookie.
        user:
          $ref: '#/components/schemas/UserSummary'
    TokenResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string

    # --- Users / Profile (REQ-003, REQ-004) ----------------------------------
    UserSummary:
      type: object
      required: [id, email, role, profile]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
        role:
          type: string
          enum: [athlete, coach]
        profile:
          oneOf:
            - $ref: '#/components/schemas/AthleteProfile'
            - $ref: '#/components/schemas/CoachProfile'
          discriminator:
            propertyName: role
    AthleteProfile:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          example: "Jane Lifter"
        weightClass:
          type: string
          nullable: true
          example: "72kg"
        dateOfBirth:
          type: string
          format: date
          nullable: true
          example: "1995-06-15"
    CoachProfile:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          example: "Coach Mike"
        bio:
          type: string
          nullable: true
          example: "USAPL certified coach, 10 years experience."
...
    ExerciseSlot:
      type: object
      required: [slotId, exerciseId, order, targetSets, targetReps]
      properties:
        slotId:
          $ref: '#/components/schemas/UUID'
        exerciseId:
          $ref: '#/components/schemas/UUID'
        order:
          type: integer
          minimum: 0
          description: Zero-based display order within the day.
        targetSets:
          type: integer
          minimum: 1
        targetReps:
          type: integer
          minimum: 1
        targetWeightLbs:
          type: number
          format: float
          nullable: true
        targetPercent1RM:
          type: number
          format: float
          minimum: 0
          maximum: 100
          nullable: true
          description: Mutually exclusive with targetWeightLbs.
        targetRPE:
          type: number
          format: float
          minimum: 1
          maximum: 10
          nullable: true
    ProgramDay:
      type: object
      required: [id, programId, weekNumber, dayNumber, slots, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        programId:
          $ref: '#/components/schemas/UUID'
        weekNumber:
          type: integer
          minimum: 1
        dayNumber:
          type: integer
          minimum: 1
          maximum: 7
        label:
          type: string
          nullable: true
          example: "Squat Day"
        slots:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseSlot'
        updatedAt:
          $ref: '#/components/schemas/ISODateTime'
    ProgramDayList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProgramDay'
    SlotInput:
      type: object
      required: [exerciseId, order, targetSets, targetReps]
      properties:
        exerciseId:
          $ref: '#/components/schemas/UUID'
        order:
          type: integer
          minimum: 0
        targetSets:
          type: integer
          minimum: 1
        targetReps:
          type: integer
          minimum: 1
        targetWeightLbs:
          type: number
          format: float
          nullable: true
        targetPercent1RM:
          type: number
          format: float
          minimum: 0
          maximum: 100
          nullable: true
        targetRPE:
          type: number
          format: float
          minimum: 1
          maximum: 10
          nullable: true
    CreateProgramDayRequest:
      type: object
      required: [weekNumber, dayNumber]
      properties:
        weekNumber:
          type: integer
          minimum: 1
        dayNumber:
          type: integer
          minimum: 1
          maximum: 7
        label:
          type: string
          nullable: true
        slots:
          type: array
          items:
            $ref: '#/components/schemas/SlotInput'
          default: []
    UpdateProgramDayRequest:
      type: object
      minProperties: 1
      properties:
        label:
          type: string
          nullable: true
        slots:
          type: array
          items:
            $ref: '#/components/schemas/SlotInput'
          description: Replaces all slots when provided.

    # --- Assignments (REQ-010, REQ-011) ---------------------------------------
...
  responses:
    400BadRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/validation-error"
            title: "Validation Error"
            status: 400
            detail: "password must be at least 8 characters"
    401Unauthorized:
      description: Missing or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Bearer token is missing or expired"
    403Forbidden:
      description: Authenticated but insufficient role or resource ownership.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/forbidden"
            title: "Forbidden"
            status: 403
            detail: "You do not have access to this resource"
    404NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/not-found"
            title: "Not Found"
            status: 404
    409Conflict:
      description: Conflict with existing state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "https://api.example.com/problems/conflict"
            title: "Conflict"
            status: 409
            detail: "An account with this email already exists"
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Register a new account
      x-req: [REQ-001]
      description: |
        **REQ-001** — Creates a new user with email, password, and role.
        Rejects duplicate emails (409). Password must be ≥ 8 characters.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: Account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "409":
          $ref: '#/components/responses/409Conflict'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Log in with email and password
      x-req: [REQ-002]
      description: |
        **REQ-002** — Returns a JWT access token and sets a refresh token
        (`httpOnly` cookie). Incorrect credentials return 401 without revealing
        which field is wrong.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          $ref: '#/components/responses/401Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logoutUser
      summary: Log out and end the current session
      x-req: [REQ-002]
      description: |
        **REQ-002** — Invalidates the refresh token cookie and the server-side
        session reference.
      responses:
        "204":
          description: Logged out successfully.
        "401":
          $ref: '#/components/responses/401Unauthorized'
  /auth/token:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Exchange refresh token cookie for a new access token
      description: |
        Uses the httpOnly refresh token cookie set during login or registration.
        No request body required. Returns a new access token.
      security: []
      responses:
        "200":
          description: New access token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          $ref: '#/components/responses/401Unauthorized'

  # ---------------------------------------------------------------------------
  # USERS / PROFILE  ·  REQ-003, REQ-004
  # ---------------------------------------------------------------------------
...
  /programs/{programId}/days:
    get:
      tags: [Programs]
      operationId: listProgramDays
      summary: List all training days for a program
      x-req: [REQ-009, REQ-011]
      description: |
        **REQ-009, REQ-011** — Returns all program_day documents ordered by
        `weekNumber`, `dayNumber`. Accessible to the owning coach and to
        athletes assigned to the program.
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Program days.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramDayList'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'

    post:
      tags: [Programs]
      operationId: createProgramDay
      summary: Add a training day to a program
      x-req: [REQ-009]
      description: |
        **REQ-009** — Creates a new day with the given week/day numbers and
        optional exercise slots. Returns 409 if that week/day pair already
        exists for this program. Coach role required.
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProgramDayRequest'
      responses:
        "201":
          description: Program day created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramDay'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409Conflict'
  /programs/{programId}/days/{dayId}:
    get:
      tags: [Programs]
      operationId: getProgramDay
      summary: Get a single training day
      x-req: [REQ-009, REQ-011]
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: dayId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Program day detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramDay'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'

    patch:
      tags: [Programs]
      operationId: updateProgramDay
      summary: Update exercises or order within a training day
      x-req: [REQ-009]
      description: |
        **REQ-009** — Partial update of a program day. When `slots` is provided
        it replaces all existing slots (full replacement of the ordered list).
        Coach role, program ownership required.
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: dayId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProgramDayRequest'
      responses:
        "200":
          description: Program day updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramDay'
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401Unauthorized'
        "403":
          $ref: '#/components/responses/403Forbidden'
        "404":
          $ref: '#/components/responses/404NotFound'

  # ---------------------------------------------------------------------------
  # ASSIGNMENTS  ·  REQ-010, REQ-011
  # ---------------------------------------------------------------------------
...